+++ webapp/src/pages/dashboard/offer/cje/[id]/index.tsx
@@ -1,15 +1,17 @@
-import { Center, useDisclosure } from "@chakra-ui/react";
+import { Center, useDisclosure, useToast } from "@chakra-ui/react";
+import { AnimatePresence, motion } from "framer-motion";
 import { useRouter } from "next/router";
 import { useEffect, useState } from "react";
+import { IoCloseCircleOutline } from "react-icons/io5";
 import CouponCard from "~/components/cards/CouponCard";
 import OfferCard from "~/components/cards/OfferCard";
 import LoadingLoader from "~/components/LoadingLoader";
+import CouponContent from "~/components/offer/page/CouponContent";
+import OfferContent from "~/components/offer/page/OfferContent";
+import ToastComponent from "~/components/ToastComponent";
 import OfferHeaderWrapper from "~/components/wrappers/OfferHeaderWrapper";
 import { api } from "~/utils/api";
-import { motion, AnimatePresence } from "framer-motion";
-import OfferContent from "~/components/offer/page/OfferContent";
-import CouponContent from "~/components/offer/page/CouponContent";
-import { isIOS } from "~/utils/tools";
+import { dateDiffInMinutes, isIOS, isOlderThan24Hours } from "~/utils/tools";
 
 const flipVariants = {
   hidden: { rotateY: 90 },
@@ -19,6 +21,9 @@ const flipVariants = {
 
 export default function OfferCjePage() {
   const router = useRouter();
+  const toast = useToast();
+
+  const [couponErased, setCouponErased] = useState(false);
 
   const { id: offer_id } = router.query as { id: string };
 
@@ -40,6 +45,35 @@ export default function OfferCjePage() {
   const { data: offer } = resultOffer || {};
   const { data: coupon } = resultCoupon || {};
 
+  // There is 3 ways user see the "Voir mon code" button
+  //		1 - User does not have a coupon assigned
+  //		2 - User has an unused coupon assigned
+  //		3 - User has a used coupon assigned but offer is cumulative
+  const canTakeCoupon =
+    !coupon ||
+    (!!coupon && !coupon.used) ||
+    (!!coupon && !!coupon.used && !!offer?.cumulative);
+
+  const hasUnusedCoupon = !!coupon && !coupon.used;
+
+  // Non cumulative offer
+  const disabled = !offer?.cumulative && !!coupon && !!coupon.used;
+
+  // Cumulative offer
+  const cooldownInMinutes =
+    !!offer?.cumulative &&
+    !!coupon &&
+    !!coupon.used &&
+    !!coupon.assignUserAt &&
+    !isOlderThan24Hours(coupon.assignUserAt)
+      ? dateDiffInMinutes(
+          new Date(),
+          new Date(
+            new Date(coupon.assignUserAt).getTime() + 24 * 60 * 60 * 1000
+          )
+        )
+      : null;
+
   const { mutateAsync: increaseNbSeen } =
     api.offer.increaseNbSeen.useMutation();
 
@@ -47,7 +81,27 @@ export default function OfferCjePage() {
     mutateAsync: mutateAsyncCouponToUser,
     isLoading: isLoadingCouponToUser,
   } = api.coupon.assignToUser.useMutation({
-    onSuccess: () => refetchCoupon(),
+    onSuccess: () => {
+      refetchCoupon();
+    },
+    onError: (error) => {
+      if (error.data?.httpStatus === 404) {
+        toast({
+          render: () => (
+            <ToastComponent
+              bgColor="error"
+              text="Plus aucun code disponible pour cette offre, redirection..."
+              icon={IoCloseCircleOutline}
+            />
+          ),
+          duration: 2000,
+        });
+
+        setTimeout(() => {
+          router.push("/dashboard");
+        }, 2000);
+      }
+    },
   });
 
   const [isFirstLoad, setIsFirstLoad] = useState(true);
@@ -59,9 +113,16 @@ export default function OfferCjePage() {
     offerId: number,
     displayCoupon: boolean = true
   ) => {
-    if (!coupon) await mutateAsyncCouponToUser({ offer_id: offerId });
-    else if (coupon && coupon.used) return;
-    if (displayCoupon) setKind("coupon");
+    if (hasUnusedCoupon) {
+      setKind("coupon");
+    } else if (canTakeCoupon) {
+      try {
+        await mutateAsyncCouponToUser({ offer_id: offerId });
+        if (displayCoupon) setKind("coupon");
+      } catch (e) {
+        console.error(e);
+      }
+    }
   };
 
   const handleBookmarkOfferToUser = async () => {
@@ -77,8 +138,11 @@ export default function OfferCjePage() {
   const [timeoutProgress, setTimeoutProgress] = useState<number>(0);
 
   const onCouponUsed = () => {
-    refetchCoupon();
-    setKind("offer");
+    setCouponErased(true);
+    setTimeout(() => {
+      refetchCoupon();
+      setKind("offer");
+    }, 1200);
   };
 
   const {
@@ -151,6 +215,12 @@ export default function OfferCjePage() {
     }
   }, [router.isReady, isLoadingCoupon]);
 
+  useEffect(() => {
+    if (kind === "offer") {
+      setCouponErased(false);
+    }
+  }, [kind]);
+
   if (isLoadingOffer || !offer || isLoadingCoupon || !router.isReady)
     return (
       <OfferHeaderWrapper
@@ -193,6 +263,7 @@ export default function OfferCjePage() {
             ) : (
               <motion.div style={{ backfaceVisibility: "hidden" }} layout>
                 <CouponCard
+                  erased={couponErased}
                   coupon={coupon}
                   handleOpenExternalLink={onOpenExternalLink}
                 />
@@ -209,7 +280,9 @@ export default function OfferCjePage() {
           offer={offer}
           handleValidateOffer={handleValidateOffer}
           isLoadingValidateOffer={isLoadingCouponToUser}
-          coupon={coupon}
+          canTakeCoupon={canTakeCoupon}
+          cooldownInMinutes={cooldownInMinutes}
+          disabled={disabled}
         />
       ) : (
         coupon && (